app/Python/requirements.txt
# pip install -r requirements.txt

requests
beautifulsoup4
gspread
pandas
pytesseract
pdfplumber
pillow
selenium
webdriver_manager
google-cloud-vision
google-cloud-storage
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client
google-ads
playwright
mysql-connector-python
python-dotenv
anthropic
openai
google-analytics-data
opencv-python
numpy
rembg
onnxruntime
rapidfuzz
fuzzywuzzy
python-Levenshtein